+++ hlp/chat/.gitignore
@@ -9,7 +9,8 @@ __pycache__/
 # Distribution / packaging
 .Python
 #data/
-free/checkpoint/
+free/pytorch_src/checkpoint/
+free/tensorflow_src/checkpoint/
 task/checkpoint/
 build/
 develop-eggs/
