--- hlp/chat/free/config/config.json
@@ -1,51 +1,0 @@
-{
-  "seq2seq": {
-    "vocab_size": 300,
-    "embedding_dim": 256,
-    "units": 1024,
-    "max_train_data_size": 36,
-    "max_length": 40
-  },
-  "transformer": {
-    "num_layers": 2,
-    "d_model": 256,
-    "num_heads": 8,
-    "units": 512,
-    "dropout": 0.1,
-    "vocab_size": 300,
-    "embedding_dim": 256,
-    "max_train_data_size": 36,
-    "max_length": 40
-  },
-  "smn": {
-    "embedding_dim": 200,
-    "max_sentence": 50,
-    "max_utterance": 10,
-    "units": 200,
-    "vocab_size": 2000,
-    "max_train_data_size": 36,
-    "learning_rate": 0.001,
-    "max_valid_data_size": 100,
-    "max_database_size": 1700
-  },
-  "seq2seq_dict_fn": "data\\seq2seq_dict.json",
-  "seq2seq_checkpoint": "checkpoint\\seq2seq",
-  "transformer_dict_fn": "data\\transformer_dict.json",
-  "transformer_checkpoint": "checkpoint\\transformer",
-  "smn_dict_fn": "data\\smn_dict_fn",
-  "smn_checkpoint": "checkpoint\\smn",
-  "resource_data": "data\\train.txt",
-  "tokenized_data": "data\\train_tokenized.txt",
-  "lccc_data": "data\\LCCC.json",
-  "lccc_tokenized_data": "data\\lccc_tokenized.txt",
-  "douban_tokenized_data": "data\\douban.txt",
-  "ubuntu_tokenized_data": "data\\ubuntu_train.txt",
-  "ubuntu_valid_data": "data\\ubuntu_valid.txt",
-  "candidate_database": "data\\candidate.json",
-  "batch_size": 32,
-  "buffer_size": 20000,
-  "beam_size": 3,
-  "epochs": 10,
-  "start_sign": "start",
-  "end_sign": "end"
-}
