--- hlp/chat/free/pytorch/config/config.json
@@ -1,53 +1,0 @@
-{
-  "seq2seq": {
-    "vocab_size": 300,
-    "embedding_dim": 256,
-    "units": 1024,
-    "max_train_data_size": 5,
-    "max_length": 40,
-    "dropout": 0.5
-  },
-  "transformer": {
-    "num_layers": 2,
-    "d_model": 256,
-    "num_heads": 8,
-    "units": 512,
-    "dropout": 0.1,
-    "vocab_size": 300,
-    "embedding_dim": 256,
-    "max_train_data_size": 36,
-    "max_length": 40
-  },
-  "smn": {
-    "embedding_dim": 200,
-    "max_sentence": 50,
-    "max_utterance": 10,
-    "units": 200,
-    "vocab_size": 2000,
-    "max_train_data_size": 36,
-    "learning_rate": 0.001,
-    "max_valid_data_size": 100,
-    "max_database_size": 1700
-  },
-  "seq2seq_dict_fn": "pytorch\\data\\seq2seq_dict.json",
-  "seq2seq_checkpoint": "pytorch\\checkpoint\\seq2seq",
-  "transformer_dict_fn": "pytorch\\data\\transformer_dict.json",
-  "transformer_checkpoint": "pytorch\\checkpoint\\transformer",
-  "smn_dict_fn": "pytorch\\data\\smn_dict_fn",
-  "smn_checkpoint": "pytorch\\checkpoint\\smn",
-  "resource_data": "data\\train.txt",
-  "tokenized_data": "pytorch\\data\\train_tokenized.txt",
-  "lccc_data": "data\\LCCC.json",
-  "lccc_tokenized_data": "pytorch\\data\\lccc_tokenized.txt",
-  "douban_tokenized_data": "data\\douban.txt",
-  "ubuntu_tokenized_data": "data\\ubuntu_train.txt",
-  "ubuntu_valid_data": "data\\ubuntu_valid.txt",
-  "candidate_database": "pytorch\\data\\candidate.json",
-  "batch_size": 2,
-  "buffer_size": 20000,
-  "beam_size": 3,
-  "epochs": 2,
-  "start_sign": "<sos>",
-  "end_sign": "<eoe>",
-  "unk_sign": "<unk>"
-}
