--- hlp/chat/free/pytorch/data/seq2seq_dict.json
@@ -1,164 +1,0 @@
-{
-    "<pad>": 0,
-    "0": "<pad>",
-    "<unk>": 1,
-    "1": "<unk>",
-    "</s>": 2,
-    "2": "</s>",
-    "<s>": 3,
-    "3": "<s>",
-    "<copy>": 4,
-    "4": "<copy>",
-    "<sos>": 5,
-    "5": "<sos>",
-    "你": 6,
-    "6": "你",
-    "去": 7,
-    "7": "去",
-    "那儿": 8,
-    "8": "那儿",
-    "竟然": 9,
-    "9": "竟然",
-    "不喊": 10,
-    "10": "不喊",
-    "我": 11,
-    "11": "我",
-    "生气": 12,
-    "12": "生气",
-    "了": 13,
-    "13": "了",
-    "，": 14,
-    "14": "，",
-    "快点": 15,
-    "15": "快点",
-    "给": 16,
-    "16": "给",
-    "道歉": 17,
-    "17": "道歉",
-    "<eoe>": 18,
-    "18": "<eoe>",
-    "！": 19,
-    "19": "！",
-    "再有": 20,
-    "20": "再有",
-    "时间": 21,
-    "21": "时间",
-    "找": 22,
-    "22": "找",
-    "我用": 23,
-    "23": "我用",
-    "SEED.24": 24,
-    "24": "SEED.24",
-    "小时": 25,
-    "25": "小时",
-    "签到": 26,
-    "26": "签到",
-    "一次": 27,
-    "27": "一次",
-    "可以": 28,
-    "28": "可以",
-    "用": 29,
-    "29": "用",
-    "4": 30,
-    "30": "4",
-    "对于": 31,
-    "31": "对于",
-    "这种": 32,
-    "32": "这种",
-    "每天晚上": 33,
-    "33": "每天晚上",
-    "逛": 34,
-    "34": "逛",
-    "一下": 35,
-    "35": "一下",
-    "的": 36,
-    "36": "的",
-    "感觉": 37,
-    "37": "感觉",
-    "不错": 38,
-    "38": "不错",
-    "SEED": 39,
-    "39": "SEED",
-    "早上": 40,
-    "40": "早上",
-    "刚": 41,
-    "41": "刚",
-    "被": 42,
-    "42": "被",
-    "禁用": 43,
-    "43": "禁用",
-    "还有": 44,
-    "44": "还有",
-    "一个月": 45,
-    "45": "一个月",
-    "VIP": 46,
-    "46": "VIP",
-    "路线": 47,
-    "47": "路线",
-    "呢": 48,
-    "48": "呢",
-    "禁": 49,
-    "49": "禁",
-    "之后": 50,
-    "50": "之后",
-    "才": 51,
-    "51": "才",
-    "买": 52,
-    "52": "买",
-    "另": 53,
-    "53": "另",
-    "一个": 54,
-    "54": "一个",
-    "一年": 55,
-    "55": "一年",
-    "结果": 56,
-    "56": "结果",
-    "一": 57,
-    "57": "一",
-    "下午": 58,
-    "58": "下午",
-    "就": 59,
-    "59": "就",
-    "挂": 60,
-    "60": "挂",
-    "现在": 61,
-    "61": "现在",
-    "个": 62,
-    "62": "个",
-    "极速网": 63,
-    "63": "极速网",
-    "速差": 64,
-    "64": "速差",
-    "很": 65,
-    "65": "很",
-    "咬咬牙": 66,
-    "66": "咬咬牙",
-    "这回": 67,
-    "67": "这回",
-    "要": 68,
-    "68": "要",
-    "全入": 69,
-    "69": "全入",
-    "领个": 70,
-    "70": "领个",
-    "搓衣板": 71,
-    "71": "搓衣板",
-    "吧": 72,
-    "72": "吧",
-    "心疼": 73,
-    "73": "心疼",
-    "干完": 74,
-    "74": "干完",
-    "这": 75,
-    "75": "这",
-    "一票": 76,
-    "76": "一票",
-    "会员": 77,
-    "77": "会员",
-    "等级": 78,
-    "78": "等级",
-    "就要": 79,
-    "79": "就要",
-    "升": 80,
-    "80": "升"
-}
