--- hlp/chat/free/pytorch/model/transformer.py
@@ -1,2 +1,0 @@
-import torch.nn as nn
-
