+++ hlp/chat/free/tensorflow_src/model/seq2seq.py
@@ -1,0 +1,59 @@
+import tensorflow as tf
+import utils.layers as layers
+
+
+class Encoder(tf.keras.Model):
+    """
+    seq2seq的encoder，主要就是使用Embedding和GRU对输入进行编码，
+    这里需要注意传入一个初始化的隐藏层，随机也可以，但是我这里就
+    直接写了一个隐藏层方法。
+    """
+
+    def __init__(self, vocab_size, embedding_dim, enc_units, batch_sz):
+        super(Encoder, self).__init__()
+        self.batch_sz = batch_sz
+        self.enc_units = enc_units
+        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
+        self.gru = tf.keras.layers.GRU(self.enc_units,
+                                       return_sequences=True,
+                                       return_state=True,
+                                       recurrent_initializer='glorot_uniform')
+
+    def call(self, x, hidden):
+        x = self.embedding(x)
+        output, state = self.gru(x, initial_state=hidden)
+        return output, state
+
+    def initialize_hidden_state(self):
+        return tf.zeros((self.batch_sz, self.enc_units))
+
+
+class Decoder(tf.keras.Model):
+    """
+    seq2seq的decoder，将初始化的x、隐藏层和encoder的输出作为
+    输入，encoder的输入用来和隐藏层进行attention，得到的上下文
+    向量和x进行整合然后丢到gru里去，最后Dense输出一下
+    """
+
+    def __init__(self, vocab_size, embedding_dim, dec_units, batch_sz):
+        super(Decoder, self).__init__()
+        self.batch_sz = batch_sz
+        self.dec_units = dec_units
+        self.embedding = tf.keras.layers.Embedding(vocab_size, embedding_dim)
+        self.gru = tf.keras.layers.GRU(self.dec_units,
+                                       return_sequences=True,
+                                       return_state=True,
+                                       recurrent_initializer='glorot_uniform')
+        self.fc = tf.keras.layers.Dense(vocab_size)
+        self.attention = layers.BahdanauAttention(self.dec_units)
+
+    def call(self, x, hidden, enc_output):
+        context_vector, attention_weights = self.attention(hidden, enc_output)
+        x = self.embedding(x)
+        x = tf.concat([tf.expand_dims(context_vector, 1), x], axis=-1)
+        output, state = self.gru(x)
+
+        output = tf.reshape(output, (-1, output.shape[2]))
+        x = self.fc(output)
+
+        return x, state, attention_weights
