--- hlp/chat/requirements.txt
@@ -1,6 +1,0 @@
-tensorflow=2.3
-torch=1.7.0
-torchtext=0.8.0
-zhon.hanzi=1.1.5
-jieba=0.42.1
-nltk=3.5
